Separar clases y solicitudes en panel tutor
